PluginBase methods moved to PluginHelper
